refactor: 금액 관련 계산에 BigDecimal 적용

+++ src/main/java/blackjack/domain/gamer/Player.java
@@ -23,7 +23,7 @@ public final class Player extends Gamer {
     }
 
     public int calculateProfit(final GameResult gameResult) {
-        return bettingMoney.calculateProfit(gameResult);
+        return bettingMoney.calculateProfit(gameResult).intValue();
     }
 
     public String getName() {
